book-project/requirements.txt
Click==7.0
Flask==1.0.2
Flask-Sockets==0.2.1
gevent==1.3.7
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
pkg-resources==0.0.0
Werkzeug==0.14.1
pandas
django-encrypted-model-fields